Revert "Feature/loss delta"
