+++ .gitignore
@@ -1,3 +1,1 @@
 *.pyc
-*.DS_Store
-covid19-library/src/mlflow_credentials.json
