+++ covid19-library/src/entities/metric_name.py
@@ -5,7 +5,3 @@ class MetricName(str, enum.Enum):
     rmse = "rmse"
     rmsle = "rmsle"
     mape = "mape"
-    rmse_delta = "rmse_delta"
-    rmsle_delta = "rmsle_delta"
-    mape_delta = "mape_delta"
-
