+++ covid19-library/src/forecast_config.json
@@ -1,6 +1,5 @@
 {
-	"data_source": "",
-	"region_name": [],
+	"region_name": "",
 	"region_type": "",
 	"run_day": "",
 	"forecast_start_date": "",
