+++ covid19-library/src/nb_utils.py
@@ -1,7 +1,6 @@
 import json
 import pandas as pd
 import matplotlib.pyplot as plt
-import mlflow
 
 from copy import deepcopy
 from datetime import datetime, timedelta
@@ -49,6 +48,8 @@ def parse_params(parameters, interval='Train1'):
     """
     param_dict = dict() # The flattened dictionary to return
     for param in parameters:
+        if 'All' in param:
+            continue
         if isinstance(parameters[param], dict):
             for key in parameters[param]:
                 assert (not isinstance(parameters[param][key], dict))
@@ -57,22 +58,6 @@ def parse_params(parameters, interval='Train1'):
         else:
             param_dict[interval + '_' + param] = parameters[param]
     return param_dict
-
-def parse_metrics(metrics, interval = 'Train1'):
-    """
-        Flatten the list of loss metrics to enable logging.
-    
-    Note:
-        Consider only losses computed on individual buckets.
-    """
-    metric_dict = dict()
-    for metric in metrics:
-        prefix = interval + '_' + metric['metric_name'] + '_'
-        variables = metric['variable_weights']
-        if len(variables) == 1:
-            metric_dict[prefix + variables[0]['variable']] = metric['value']
-            metric_dict[prefix + variables[0]['variable'] + '_weight'] = variables[0]['weight']
-    return metric_dict
 
 
 def train_eval(region, region_type, 
@@ -85,7 +70,7 @@ def train_eval(region, region_type,
     """
         #TODO: Need to add hooks to consume data from appropriate source
 
-        Run train and evualation for (basic) SEIR model.
+        Run train and evalation for (basic) SEIR model.
     
     Arguments:
         region, region_type : Region info corresponding to the run
@@ -117,7 +102,7 @@ def train_eval(region, region_type,
     params = dict()
     metrics = dict()
 
-    params['region'] = " ".join(region)
+    params['region'] = region
     params['region_type'] = region_type
     params['train1_start_date'] = train1_start_date
     params['train1_end_date'] = train1_end_date
@@ -129,7 +114,6 @@ def train_eval(region, region_type,
     params['data_source'] = data_source
 
     train_config = deepcopy(default_train_config)
-    train_config['data_source'] = data_source
     train_config['region_name'] = region
     train_config['region_type'] = region_type
     train_config['train_start_date'] = train1_start_date
@@ -138,7 +122,6 @@ def train_eval(region, region_type,
     
     # model parameters
     model_params = dict()
-    model_params['data_source'] = data_source
     model_params['region'] = region
     model_params['region_type'] = region_type
     model_params['model_type'] = train_config['model_class']
@@ -165,13 +148,12 @@ def train_eval(region, region_type,
 
     metrics['Train1RMLSE'] = train1RMSLE
     metrics['Train1MAPE'] = train1MAPE
+    metrics['Train1All'] = trainResults['train_metric_results']
     metrics.update(parse_params(trainResults['best_params'], 'Train1'))
     metrics.update(parse_params(trainResults['latent_params'], 'Train1')) 
-    metrics.update(parse_metrics(trainResults['train_metric_results'], 'Train1')) 
     train1_model_params['model_parameters'] = trainResults['model_parameters']
     
     test_config = deepcopy(default_test_config)
-    test_config['data_source'] = data_source
     test_config['region_name'] = region
     test_config['region_type'] = region_type
     test_config['test_start_date'] = test_start_date
@@ -198,10 +180,9 @@ def train_eval(region, region_type,
 
     metrics['TestMAPE'] = testMAPE
     metrics['TestRMLSE'] = testRMSLE
-    metrics.update(parse_metrics(evalResults, 'Test')) 
+    metrics['TestAll'] = evalResults
     
     finalTrain_config = deepcopy(default_train_config)
-    finalTrain_config['data_source'] = data_source
     finalTrain_config['region_name'] = region
     finalTrain_config['region_type'] = region_type
     finalTrain_config['train_start_date'] = train2_start_date
@@ -227,10 +208,10 @@ def train_eval(region, region_type,
 
     metrics['Train2MAPE'] = train2MAPE
     metrics['Train2RMLSE'] = train2RMSLE
+    metrics['Train2All'] = finalResults['train_metric_results']
     metrics.update(parse_params(finalResults['best_params'], 'Train2'))
     metrics.update(parse_params(finalResults['latent_params'], 'Train2'))
-    metrics.update(parse_metrics(finalResults['train_metric_results'], 'Train2')) 
-    
+        
     model_params['model_parameters'] = finalResults['model_parameters']
     train2_model_params['model_parameters'] = finalResults['model_parameters']
     
@@ -251,7 +232,6 @@ def forecast(model_params, run_day, forecast_start_date, forecast_end_date,
         forecast_df : Dataframe containing forecasts
     """
     evalConfig = ForecastingModuleConfig.parse_obj(default_forecast_config)
-    evalConfig.data_source = model_params['data_source']
     evalConfig.region_name = model_params['region']
     evalConfig.region_type = model_params['region_type']
     evalConfig.model_parameters = model_params['model_parameters']
@@ -269,14 +249,14 @@ def forecast(model_params, run_day, forecast_start_date, forecast_end_date,
     return forecast_df
 
 
-def get_observations_in_range(data_source, region_name, region_type, 
+def get_observations_in_range(region_name, region_type, 
                               start_date, end_date,
                               obs_type = 'confirmed'):
     """
         Return a list of counts of obs_type cases
         from the region in the specified date range.
     """
-    observations = DataFetcherModule.get_observations_for_region(region_type, region_name, data_source)
+    observations = DataFetcherModule.get_observations_for_region(region_type, region_name)
     observations_df = observations[observations['observation'] == obs_type]
     
     start_date = datetime.strptime(start_date, '%m/%d/%y')
@@ -312,10 +292,9 @@ def plot(model_params, forecast_df, forecast_start_date, forecast_end_date, plot
     assert end_date < datetime.now()    
     
     # Fetch actual counts from the DataFetcher module
-    data_source = model_params['data_source']
     region_name = model_params['region']
     region_type = model_params['region_type']
-    actual_observations = DataFetcherModule.get_observations_for_region(region_name, region_type, data_source)
+    actual_observations = DataFetcherModule.get_observations_for_region(region_name, region_type)
     
     # Get relevant time-series of actual counts from actual_observations
     actual_observations = get_observations_in_range(region_name, region_type, 
@@ -373,13 +352,6 @@ def plot_m1(train1_model_params, train1_run_day, train1_start_date, train1_end_d
             rolling_average = False, uncertainty = False, 
             forecast_config = 'forecast_config.json',
             plot_config = 'plot_config.json', plot_name = 'default.png'):
-    """
-        M1 plot consisting of:
-            - Actuals for train1 and test intervals
-            - Rolling average for train1 and test intervals
-            - M1 predictions for train1 period initialized on train1 run day
-            - M1 predictions for test period initialized on test run day
-    """
     
     ## TODO: Log scale
     with open(plot_config) as fin:
@@ -393,6 +365,7 @@ def plot_m1(train1_model_params, train1_run_day, train1_start_date, train1_end_d
 
     with open(forecast_config) as fin:
         default_forecast_config = json.load(fin)
+
     
     # Get predictions
     pd_df_train = forecast(train1_model_params, train1_run_day, train1_start_date, train1_end_date, default_forecast_config)
@@ -404,8 +377,7 @@ def plot_m1(train1_model_params, train1_run_day, train1_start_date, train1_end_d
     pd_df_test = pd_df_test.sort_values(by=['index'])
 
     # Get observed data
-    actual = DataFetcherModule.get_observations_for_region(
-        train1_model_params['region_type'], train1_model_params['region'], train1_model_params['data_source'])
+    actual = DataFetcherModule.get_observations_for_region(train1_model_params['region_type'], train1_model_params['region'])
     actual = actual.set_index('observation')
     actual = actual.transpose()
     actual = actual.reset_index()
@@ -475,16 +447,10 @@ def plot_m2(train2_model_params, train_start_date, train_end_date,
             rolling_average = False, uncertainty = False, 
             forecast_config = 'forecast_config.json',
             plot_config = 'plot_config.json', plot_name = 'default.png'):
-    """
-        M2 plot consisting of:
-            - Actuals for train2 interval and preceding weeks
-            - Rolling average for train2 interval and preceding weeks
-            - M2 predictions for train2 period initialized on train2 run day
-    """
     
     ## TODO: Log scale
     with open(plot_config) as fplot, \
-    open(forecast_config) as fcast:
+        open(forecast_config) as fcast:
         default_plot_config = json.load(fplot)
         default_forecast_config = json.load(fcast)
     
@@ -501,8 +467,7 @@ def plot_m2(train2_model_params, train_start_date, train_end_date,
     pd_df = pd_df_test.sort_values(by=['index'])
 
     # Get observed data
-    actual = DataFetcherModule.get_observations_for_region(
-        train2_model_params['region_type'], train2_model_params['region'], train2_model_params['data_source'])
+    actual = DataFetcherModule.get_observations_for_region(train2_model_params['region_type'], train2_model_params['region'])
     actual = actual.set_index('observation')
     actual = actual.transpose()
     actual = actual.reset_index()
@@ -554,7 +519,6 @@ def plot_m2(train2_model_params, train_start_date, train_end_date,
     ax.xaxis.set_minor_locator(mdates.DayLocator(interval=1))
     ax.xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))
     
-    plt.title(train2_model_params['region'])
     plt.ylabel('No of People')
     plt.xlabel('Time')
     plt.legend()
@@ -568,14 +532,8 @@ def plot_m3(train2_model_params, train1_start_date,
             rolling_average = False, uncertainty = False,
             forecast_config = 'forecast_config.json',
             plot_config = 'plot_config.json', plot_name = 'default.png'):
-    """
-        M3 plot consisting of:
-            - Forecast from forecast_start_date
-            - Actuals for preceding weeks
-    """
     
     ## TODO: Log scale
-    
     with open(plot_config) as fplot, \
         open(forecast_config) as fcast:
         default_plot_config = json.load(fplot)
@@ -597,8 +555,7 @@ def plot_m3(train2_model_params, train1_start_date,
     pd_df = pd_df_forecast.sort_values(by=['index'])
 
     # Get observed data
-    actual = DataFetcherModule.get_observations_for_region(
-        train2_model_params['region_type'], train2_model_params['region'], train2_model_params['data_source'])
+    actual = DataFetcherModule.get_observations_for_region(train2_model_params['region_type'], train2_model_params['region'])
     actual = actual.set_index('observation')
     actual = actual.transpose()
     actual = actual.reset_index()
@@ -646,119 +603,9 @@ def plot_m3(train2_model_params, train1_start_date,
     ax.xaxis.set_minor_locator(mdates.DayLocator(interval=1))
     ax.xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))
     
-    plt.title(train2_model_params['region'])
     plt.ylabel('No of People')
     plt.xlabel('Time')
     plt.legend()
     plt.grid()
     
     plt.savefig(plot_name)
-
-def set_dates(current_day):
-    
-    dates = dict()
-    
-    # Train1 : 1 week interval starting 2 weeks prior to current day
-    train1_start_date = current_day - timedelta(14)
-    train1_end_date = current_day - timedelta(8)
-    train1_run_day = train1_start_date - timedelta(1)
-
-    dates['train1_start_date'] = train1_start_date.strftime("%-m/%-d/%y")
-    dates['train1_end_date'] = train1_end_date.strftime("%-m/%-d/%y")
-    dates['train1_run_day'] = train1_run_day.strftime("%-m/%-d/%y")
-
-    # Train2 : 1 week interval starting 1 week prior to current day
-    train2_start_date = current_day - timedelta(7)
-    train2_end_date = current_day
-    train2_run_day = train2_start_date - timedelta(1)
-
-    dates['train2_start_date'] = train2_start_date.strftime("%-m/%-d/%y")
-    dates['train2_end_date'] = train2_end_date.strftime("%-m/%-d/%y")
-    dates['train2_run_day'] = train2_run_day.strftime("%-m/%-d/%y")
-
-    # Test: 1 week interval prior to current day
-    test_start_date = current_day - timedelta(7)
-    test_end_date = current_day
-    test_run_day = test_start_date - timedelta(1)
-
-    dates['test_start_date'] = test_start_date.strftime("%-m/%-d/%y")
-    dates['test_end_date'] = test_end_date.strftime("%-m/%-d/%y")
-    dates['test_run_day'] = test_run_day.strftime("%-m/%-d/%y")
-    
-    return dates
-    
-def train_eval_plot(region, region_type, 
-                    current_day, forecast_length,
-                    default_train_config, default_test_config,
-                    max_evals = 1000, data_source = None,
-                    mlflow_log = False, mlflow_run_name = None):
-    
-    """
-        Run train, evaluation and plotting. 
-    
-    Arguments:
-        region, region_type : Region info corresponding to the run
-        current_day : Indicates end of test/train2 interval
-        forecast_length : Length of forecast interval
-        default_train_config : Default train config (loaded from train_config.json)
-        default_test_config : Default test config (loaded from test_config.json)
-        max_evals : number of search evaluations for SEIR (default: 1000)
-        data_source : Data source for picking the region data
-        mlflow_log : Experiment logged using MLFlow (default: True)
-        mlflow_run_name : Name given to run in MLFlow (default: None)
-
-    Note:
-        date_format : datetime.date object 
-    """
-    
-    if not isinstance(region, list):
-        region = [region]
-        
-    name_prefix = " ".join(region)
-        
-    dates = set_dates(current_day)
-    
-    train1_start_date = dates['train1_start_date']
-    train1_end_date = dates['train1_end_date']
-    train1_run_day = dates['train1_run_day']
-    
-    train2_start_date = dates['train2_start_date']
-    train2_end_date = dates['train2_end_date']
-    train2_run_day = dates['train2_run_day']
-    
-    test_start_date = dates['test_start_date']
-    test_end_date = dates['test_end_date']
-    test_run_day = dates['test_run_day']
-    
-    with mlflow.start_run(run_name = mlflow_run_name):
-    
-        params, metrics, train1_params, train2_params = train_eval(region, region_type, 
-                                                               train1_start_date, train1_end_date, 
-                                                               train2_start_date, train2_end_date, train2_run_day,
-                                                               test_start_date, test_end_date,
-                                                               default_train_config, default_test_config,
-                                                               max_evals = max_evals, data_source = data_source, 
-                                                               mlflow_log = mlflow_log,
-                                                               name_prefix = name_prefix)
-
-        plot_m1(train1_params, train1_run_day, train1_start_date, train1_end_date, 
-            test_run_day, test_start_date, test_end_date, 
-            rolling_average = False, uncertainty = False, 
-            plot_config = 'plot_config.json', plot_name = name_prefix+'_m1.png')
-
-        plot_m2(train2_params, train1_start_date, train1_end_date, 
-            test_run_day, test_start_date, test_end_date, plot_name = name_prefix+ '_m2.png')
-
-        forecast_start_date = (datetime.strptime(train2_end_date, "%m/%d/%y") + timedelta(1)).strftime("%-m/%-d/%y")
-        plot_m3(train2_params, train1_start_date, 
-                forecast_start_date, forecast_length, plot_name = name_prefix +'_m3.png')
-        
-        mlflow.log_params(params)
-        mlflow.log_metrics(metrics)
-        mlflow.log_artifact(name_prefix+'_m1.png')
-        mlflow.log_artifact(name_prefix+'_m2.png')
-        mlflow.log_artifact(name_prefix+'_m3.png')
-        mlflow.log_artifact('train_config.json')
-        mlflow.log_artifact('train1_output.json')
-        mlflow.log_artifact('test_output.json')
-        mlflow.log_artifact('train2_output.json')
