+++ covid19-library/src/test_config.json
@@ -1,6 +1,5 @@
 {
-	"data_source": "",
-	"region_name": [],
+	"region_name": "",
 	"region_type": "",
 	"run_day": "",
 	"test_start_date": "",
