+++ covid19-library/src/train_config.json
@@ -1,6 +1,5 @@
 {
-	"data_source": "",
-	"region_name": [],
+	"region_name": "",
 	"region_type": "",
 	"train_start_date": "",
 	"train_end_date": "",
